refactor: Move paas/gitea package to internal/gitea
because it is carrier specific and shouldn't be consumed by external
projects. Also "paas" doesn't add any useful information.
